[react-hw8] Fixes after code review
