+++ react-2021-09/game-of-life/src/components/AppRouter/AppRouter.tsx
@@ -1,23 +1,43 @@
 import React from 'react';
 import { HashRouter, Route, Switch, Redirect } from 'react-router-dom';
-import { AppUI, AppUIProps } from '@src/components/AppUI';
+import { GameUI } from '@src/components/GameUI';
+import { AuthorizedHead } from '@src/components/AuthorizedHead';
+import { LoginForm } from '@src/components/LoginForm';
+import { GameSettings } from '@src/components/GameSettings';
+import { AppState } from '@src/components/AppStateManager/appReducer';
 
-export const AppRouter: React.FC<AppUIProps> = (props) => {
+interface AppRouterProps {
+    appState: AppState;
+    invert: (num: number) => void;
+    setSmall: () => void;
+    setMedium: () => void;
+    setLarge: () => void;
+    clear: () => void;
+    fill25: () => void;
+    fill50: () => void;
+    fill75: () => void;
+    fill100: () => void;
+    onChangeName: (name: string) => void;
+    onLogout: () => void;
+}
+
+export const AppRouter: React.FC<AppRouterProps> = ({ appState, ...func }) => {
     return (
         <HashRouter>
             <Switch>
                 <Route path="/login">
-                    <AppUI {...props} />
+                    <LoginForm onChangeName={func.onChangeName} />
                 </Route>
-                <Route path="/game">
-                    <AppUI {...props} />
+                <Route path="/">
+                    <AuthorizedHead userName={appState.userName} onLogout={func.onLogout} />
+                    <GameUI appState={appState} invert={func.invert} />
+                    <GameSettings appState={appState} {...func} />
                 </Route>
                 <Route path="*">
                     <Redirect to="/login" />
                 </Route>
             </Switch>
-            {!props.appState.userName && <Redirect to="/login" />}
-            {props.appState.userName && <Redirect to="/game" />}
+            {appState.userName ? <Redirect to="/" /> : <Redirect to="/login" />}
         </HashRouter>
     );
 };
