+++ react-2021-09/game-of-life/src/components/AuthorizedHead/index.tsx
@@ -1,0 +1,1 @@
+export * from './AuthorizedHead';
