+++ react-2021-09/game-of-life/src/components/GameUI/index.tsx
@@ -1,0 +1,1 @@
+export * from './GameUI';
