+++ react-2021-09/game-of-life/src/components/LoginForm/LoginForm.spec.tsx
@@ -1,14 +1,14 @@
 import React from 'react';
-import { UserNameForm } from './UserNameForm';
+import { LoginForm } from './LoginForm';
 import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import { str } from '@src/testFramework/lib/reducer';
 
-describe('UserNameForm', () => {
+describe('LoginForm', () => {
     it('calls onName() callback when submit is clicked and name is not empty', () => {
         const mockCallBack = jest.fn();
         const rndName = str();
-        render(<UserNameForm onName={mockCallBack} userName="" onLogout={() => {}} />);
+        render(<LoginForm onChangeName={mockCallBack} />);
         const input = screen.getByLabelText('Enter your name:');
         userEvent.click(input);
         userEvent.type(input, rndName);
@@ -19,7 +19,7 @@ describe('UserNameForm', () => {
     it('sends a name into props.onName() callback when submit is clicked and name is not empty', () => {
         const mockCallBack = jest.fn();
         const rndName = str();
-        render(<UserNameForm onName={mockCallBack} userName="" onLogout={() => {}} />);
+        render(<LoginForm onChangeName={mockCallBack} />);
         const input = screen.getByLabelText('Enter your name:');
         userEvent.click(input);
         userEvent.type(input, rndName);
