+++ react-2021-09/game-of-life/src/components/common.tsx
@@ -1,0 +1,11 @@
+import styled from '@emotion/styled';
+
+export const TopControls = styled.section`
+    display: flex;
+    justify-content: center;
+`;
+
+export const TableBg = styled.div`
+    background: #e9e2d7;
+    font-family: Arial, sans-serif;
+`;
